cleaned up with duplicate comparators
