+++ app/src/main/java/com/sung/guardiannews/view/adapters/GuardianArticleListAdapter.kt
@@ -2,15 +2,15 @@ package com.sung.guardiannews.view.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sung.guardiannews.databinding.NewsSectionArticleItemBinding
 import com.sung.guardiannews.model.Article
 import com.sung.guardiannews.view.GuardianNewsCallback
+import com.sung.guardiannews.view.helpers.ArticleComparator
 
 class GuardianArticleListAdapter(private val callback: GuardianNewsCallback) :
-    ListAdapter<Article, GuardianArticleListAdapter.GuardianArticleViewHolder>(ARTICLE_COMPARATOR) {
+    ListAdapter<Article, GuardianArticleListAdapter.GuardianArticleViewHolder>(ArticleComparator) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): GuardianArticleViewHolder {
         return GuardianArticleViewHolder(
@@ -32,16 +32,6 @@ class GuardianArticleListAdapter(private val callback: GuardianNewsCallback) :
             binding.model = item
             binding.callback = callback
             binding.executePendingBindings()
-        }
-    }
-
-    companion object {
-        private val ARTICLE_COMPARATOR = object : DiffUtil.ItemCallback<Article>() {
-            override fun areItemsTheSame(oldItem: Article, newItem: Article): Boolean =
-                oldItem.apiUrl == newItem.apiUrl
-
-            override fun areContentsTheSame(oldItem: Article, newItem: Article): Boolean =
-                oldItem == newItem
         }
     }
 }
