+++ app/src/main/java/com/sung/guardiannews/view/adapters/GuardianMostViewedArticleListAdapter.kt
@@ -2,16 +2,16 @@ package com.sung.guardiannews.view.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sung.guardiannews.databinding.MostViewedArticleItemBinding
 import com.sung.guardiannews.model.Article
 import com.sung.guardiannews.view.GuardianArticleCallback
+import com.sung.guardiannews.view.helpers.ArticleComparator
 
 class GuardianMostViewedArticleListAdapter(private val callback: GuardianArticleCallback) :
     ListAdapter<Article, GuardianMostViewedArticleListAdapter.GuardianArticleViewHolder>(
-        ARTICLE_COMPARATOR
+        ArticleComparator
     ) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): GuardianArticleViewHolder {
@@ -35,16 +35,6 @@ class GuardianMostViewedArticleListAdapter(private val callback: GuardianArticle
             binding.model = item
             binding.callback = callback
             binding.executePendingBindings()
-        }
-    }
-
-    companion object {
-        private val ARTICLE_COMPARATOR = object : DiffUtil.ItemCallback<Article>() {
-            override fun areItemsTheSame(oldItem: Article, newItem: Article): Boolean =
-                oldItem.apiUrl == newItem.apiUrl
-
-            override fun areContentsTheSame(oldItem: Article, newItem: Article): Boolean =
-                oldItem == newItem
         }
     }
 }
