+++ app/src/main/java/com/sung/guardiannews/view/adapters/GuardianSectionListAdapter.kt
@@ -2,15 +2,15 @@ package com.sung.guardiannews.view.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sung.guardiannews.databinding.NewsSectionArticlesBinding
 import com.sung.guardiannews.model.Section
 import com.sung.guardiannews.view.GuardianNewsCallback
+import com.sung.guardiannews.view.helpers.SectionComparator
 
 class GuardianSectionListAdapter(private val callback: GuardianNewsCallback) :
-    ListAdapter<Section, GuardianSectionListAdapter.GuardianSectionViewHolder>(SECTION_COMPARATOR) {
+    ListAdapter<Section, GuardianSectionListAdapter.GuardianSectionViewHolder>(SectionComparator) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): GuardianSectionViewHolder {
         return GuardianSectionViewHolder(
@@ -36,16 +36,6 @@ class GuardianSectionListAdapter(private val callback: GuardianNewsCallback) :
             binding.model = item
             binding.callback = callback
             binding.executePendingBindings()
-        }
-    }
-
-    companion object {
-        private val SECTION_COMPARATOR = object : DiffUtil.ItemCallback<Section>() {
-            override fun areItemsTheSame(oldItem: Section, newItem: Section): Boolean =
-                oldItem.sectionName == newItem.sectionName
-
-            override fun areContentsTheSame(oldItem: Section, newItem: Section): Boolean =
-                oldItem == newItem
         }
     }
 }
